% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.cache.load}
\alias{ifh.cache.load}
\title{Load the given cache file into environment}
\usage{
ifh.cache.load(file)
}
\arguments{
\item{file}{file path to load}
}
\description{
This function loads the given file into the environment. Currently it
is a simple wrapper around load().
}
