% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.cache.loadable}
\alias{ifh.cache.loadable}
\title{Checks if the given cache file exists and user wants to load it}
\usage{
ifh.cache.loadable(file, prompt = "Should the cache be loaded?")
}
\arguments{
\item{file}{file path of the cache file}

\item{prompt}{Prompt to show to the user if cache exists}
}
\value{
TRUE if cache exists and user wants to load it, FALSE otherwise
}
\description{
This function checks whether the cache file exists and if so, whether
the user wants to load it or not. If the cache file doesn't exist, FALSE
is returned. If the cache exists, the user is asked whether the cache
should be loaded or not. If the user decides to load it, TRUE is returned.
Else FALSE is returned.
}
