% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.create.filename}
\alias{ifh.create.filename}
\title{Creates a valid file name form the given arguments}
\usage{
ifh.create.filename(...)
}
\arguments{
\item{...}{arguments passed down to paste(..)}
}
\value{
valid file name
}
\description{
This function creates a valid filename from the given values by
concatenation using the separator '__' and replacement of '.' with '_'.
}
