% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.create.path}
\alias{ifh.create.path}
\title{Creates a valid string path from the given arguments}
\usage{
ifh.create.path(...)
}
\arguments{
\item{...}{arguments passed down to paste(..)}
}
\value{
valid path
}
\description{
This function is a simple wrapper around paste(..., sep = '').
}
