% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.dir.create}
\alias{ifh.dir.create}
\title{Creates the given directory}
\usage{
ifh.dir.create(path)
}
\arguments{
\item{path}{path to the directory}
}
\description{
This function is currently a simple wrapper around dir.create().#'
}
