% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.dir.create_if}
\alias{ifh.dir.create_if}
\title{Creates the directory if not already present}
\usage{
ifh.dir.create_if(path)
}
\arguments{
\item{path}{path to the directory}
}
\value{
TRUE if directory was created, FALSE if directory is already present
}
\description{
This function will check if the directory already exists. If so, it
returns FALSE. If not, the directory is created and TRUE is returned.
}
