% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.dir.exists}
\alias{ifh.dir.exists}
\title{Checks if directory exists}
\usage{
ifh.dir.exists(path)
}
\arguments{
\item{path}{path to directory}
}
\value{
TRUE if directory exists, FALSE otherwise
}
\description{
This function is currently a simple wrapper around dir.exists().
}
