% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.error}
\alias{ifh.error}
\title{Prints the arguments as concatenated string and marked as error}
\usage{
ifh.error(...)
}
\arguments{
\item{...}{arguments passed down to paste(..)}
}
\description{
This function prints the given message and adds the log level prefix
'[e] ERROR:'. After the call to ifh.init(), this function is replaced by
cli_alert_danger() by the 'cli' package for prettier messages.
}
\examples{
ifh.error("Print an error message with number", 10)

}
