% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.file.exists}
\alias{ifh.file.exists}
\title{Checks if file exists with given path}
\usage{
ifh.file.exists(path)
}
\arguments{
\item{path}{path to file to check}
}
\value{
TRUE if file exists, FALSE otherwise
}
\description{
This function checks if the given file exists. It wraps the calls to
file.exists() and dir.exists() to only return TRUE if its a file and
not a directory.
}
