% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.file.get.line}
\alias{ifh.file.get.line}
\title{Retrieves the first line with the given prefix}
\usage{
ifh.file.get.line(path, prefix, comment.prefix = "##", sep = "\\t")
}
\arguments{
\item{path}{the path to the file}

\item{prefix}{the prefix to detect}

\item{comment.prefix}{the prefix of comments to ignore}

\item{sep}{the separator of table cells}
}
\value{
the first line matching the prefix (prefix removed) or none
}
\description{
This method can be used to retrieve the header of a table from a file
if the header has a special line prefix that prevents read.csv(..) from detecting
the header names.
}
