% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.file.remove}
\alias{ifh.file.remove}
\title{Removes the given file}
\usage{
ifh.file.remove(path)
}
\arguments{
\item{path}{path to file for removal}
}
\value{
TRUE if removed, FALSE if file doesn't exist
}
\description{
This function removes the given file if present. It returns TRUE
if it successfully removed the file. If the file doesn't exist,
it doesn't throw an error or warning but will return FALSE.#'
}
