% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.info}
\alias{ifh.info}
\title{Prints the arguments as concatenated string and marked as info.}
\usage{
ifh.info(...)
}
\arguments{
\item{...}{arguments passed down to paste(..)}
}
\description{
This function prints the given message and adds the log level prefix
'[i] INFO:'. After the call to ifh.init(), this function is replaced by
cli_alert_info() by the 'cli' package for prettier messages.
}
\examples{
ifh.info("Print some info with number", 10)

}
