% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.init}
\alias{ifh.init}
\title{Initializes the project by setting up logging and libraries}
\usage{
ifh.init(packages, quiet = FALSE)
}
\arguments{
\item{packages}{vector of all packages to install}

\item{quiet}{suppress all output or not}
}
\description{
This function installs 'cli', 'progress' and 'optparse' by default
for pretty output. Furthermore any list of package names provided as
'packages' are installed and loaded. Also the flag is provided to
suppress all output of the installation and load calls.#'
}
\examples{
ifh.init(c('somePackage', 'otherPackage'), quiet = TRUE)

}
