% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.install}
\alias{ifh.install}
\title{Install and load the given package using BiocManager, usable after ifh.init(..)}
\usage{
ifh.install(pkg, ..., quiet = FALSE)
}
\arguments{
\item{pkg}{package name to install}

\item{...}{arguments passed down to install(..)}

\item{quiet}{suppress all output or not}
}
\description{
This function allows the installation of a package. It is only callable
without an error after ifh.init() is called. It can be used to install and load
additional packages after the ifh.init() was called.
}
\examples{
ifh.install('cli', quiet = FALSE)

}
