% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.message}
\alias{ifh.message}
\title{Print a message containing the concatenated string of all arguments.}
\usage{
ifh.message(..., prefix = "")
}
\arguments{
\item{...}{arguments passed down to paste(..)}

\item{prefix}{A prefix to add to the message}
}
\description{
This function prints the given message with the given prefix if provided.
It is mainly used by ifh.info(), ifh.step(), ifh.success() and ifh.error().
}
\examples{
ifh.message("The number is", 10)
ifh.message("Add prefix to mark the message.", prefix = "[!]")

}
