% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.performance.measure}
\alias{ifh.performance.measure}
\title{Measures the performance of the given function call}
\usage{
ifh.performance.measure(func, ...)
}
\arguments{
\item{func}{the function to measure}

\item{...}{arguments passed to func}
}
\value{
Return value of func
}
\description{
This function measures the time the given function takes to be executed.
It prints the result via ifh.info(). This function can be used if you want
to determine the potential bottleneck that consumes a lot of CPU time.
}
