% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.quiet}
\alias{ifh.quiet}
\title{Disables/Enables all output by redirecting to /dev/null}
\usage{
ifh.quiet(quiet)
}
\arguments{
\item{quiet}{defines whether output is disabled or not}
}
\description{
This function sets the sink to /dev/null or resets the sink
to the default based on the boolean value of quiet.
}
