% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.run}
\alias{ifh.run}
\title{Executes the expression and adds pretty output of warnings and errors}
\usage{
ifh.run(expr)
}
\arguments{
\item{expr}{expression to execute}
}
\description{
This function wraps the given expression and adds handling of warning and
error messages. Especially it uses the 'cli' package to prettify the
warning and error messages.
}
