% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.step}
\alias{ifh.step}
\title{Prints the arguments as concatenated string and marked as step.}
\usage{
ifh.step(...)
}
\arguments{
\item{...}{arguments passed down to paste(..)}
}
\description{
This function prints the given message and adds the log level prefix
'[-] STEP:'. After the call to ifh.init(), this function is replaced by
cli_alert() by the 'cli' package for prettier messages.
}
\examples{
ifh.step("Print some step with number", 10)

}
