% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.string.random}
\alias{ifh.string.random}
\title{Creates n random strings}
\usage{
ifh.string.random(n = 1)
}
\arguments{
\item{n}{number of random strings to create}
}
\value{
vector of random strings
}
\description{
This function returns a list of n random generated string. It can be useful
in cases where you have to create temporary files that shouldn't clash with
existing files.
}
