% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.success}
\alias{ifh.success}
\title{Prints the arguments as concatenated string and marked as success.}
\usage{
ifh.success(...)
}
\arguments{
\item{...}{arguments passed down to paste(..)}
}
\description{
This function prints the given message and adds the log level prefix
'[!] SUCCESS:'. After the call to ifh.init(), this function is replaced by
cli_alert_success() by the 'cli' package for prettier messages.
}
\examples{
ifh.success("Print a success message with number", 10)

}
