% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.table.export}
\alias{ifh.table.export}
\title{Export the table as valid csv}
\usage{
ifh.table.export(
  table,
  file,
  row.names = FALSE,
  quote = FALSE,
  sep = "\\t",
  ...
)
}
\arguments{
\item{table}{the data.frame to export}

\item{file}{the output file name}

\item{row.names}{defines whether row.names are included}

\item{quote}{defines whether quotes are included}

\item{sep}{seperator to use as cell delimiter}

\item{...}{additional values to pass down to write.table(..)}
}
\description{
This function exports the given table as valid CSV using the given separator.
It is used in cases where the write.csv() function won't work because it only allows
the comma separator.
}
