% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.table.import}
\alias{ifh.table.import}
\title{Imports the given CSV file}
\usage{
ifh.table.import(file, sep = "\\t", ...)
}
\arguments{
\item{file}{the file path to import}

\item{sep}{the separator of each cell}

\item{...}{additional parameters to pass down to read.csv(..)}
}
\value{
the imported data.frame
}
\description{
This function is currently a simple wrapper around read.csv() with additional logging.
}
