% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ifh.user.prompt}
\alias{ifh.user.prompt}
\title{Retrieves user input in (non-)interactive mode}
\usage{
ifh.user.prompt(prompt)
}
\arguments{
\item{prompt}{prompt to show to user}
}
\description{
This function prompts the user for input. This wrapper is necessary
to provide the prompt functionality in non-interactive and interactive
mode.
}
